+++ source/Messaging.Application/Transactions/MoveIn/CreateEndOfSupplyNotificationHandler.cs
@@ -16,6 +16,7 @@ using System;
 using System.Threading;
 using System.Threading.Tasks;
 using MediatR;
+using Messaging.Application.Transactions.MoveIn.Notifications;
 using Messaging.Domain.Transactions.MoveIn;
 
 namespace Messaging.Application.Transactions.MoveIn;
